+++ src/Bool.hs
@@ -5,11 +5,10 @@ module Bool
        ( whenM
        , unlessM
        , ifM
-       , guardM
        , bool
        ) where
 
-import           Control.Monad (Monad, MonadPlus, guard, unless, when, (=<<), (>>=))
+import           Control.Monad (Monad, unless, when, (>>=))
 import           Data.Bool     (Bool)
 import           Data.Function (flip)
 
@@ -27,5 +26,7 @@ unlessM p m =
 ifM :: Monad m => m Bool -> m a -> m a -> m a
 ifM p x y = p >>= \b -> if b then x else y
 
+{-
 guardM :: MonadPlus m => m Bool -> m ()
 guardM f = guard =<< f
+-}